pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "binance-sbe-feed-handler"
version = "0.1.0"
description = "Low-latency feed handler for Binance SBE Market Data Streams (spot)"
requires-python = ">=3.11"
dependencies = [
    "websockets>=13.0",
    "PyYAML>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
